+++ src/Dialect/ONNX/ONNXDimAnalysis.cpp
@@ -21,12 +21,12 @@
 #include "llvm/Support/Debug.h"
 
 #include "src/Dialect/ONNX/DialectBuilder.hpp"
+#include "src/Dialect/ONNX/ONNXDimAnalysis.hpp"
 #include "src/Dialect/ONNX/ONNXOps.hpp"
 #include "src/Dialect/ONNX/ONNXOps/OpHelper.hpp"
 #include "src/Dialect/ONNX/ONNXOps/ShapeHelper.hpp"
 #include "src/Pass/Passes.hpp"
 #include "src/Support/TypeUtilities.hpp"
-#include "src/Transform/ONNX/ONNXDimAnalysis.hpp"
 
 #define DEBUG_TYPE "dim_analysis"
