+++ src/Transform/ONNX/CMakeLists.txt
@@ -76,17 +76,3 @@ add_onnx_mlir_library(OMONNXSimplifyShapeRelatedOps
   MLIRPass
   MLIRTransforms
   )
-
-# hi alex
-# add_onnx_mlir_library(OMONNXDimAnalysis
-#   ONNXDimAnalysis.cpp
-# 
-#   LINK_LIBS PUBLIC
-#   OMONNXOps
-#   OMONNXRewrite
-#   OMShapeInference
-#   OMShapeHelperOpInterface
-#   MLIRFuncDialect
-#   MLIRPass
-#   MLIRTransforms
-#   )
